:art: 중복 구조 따로 분리
