+++ app/blueprints/standard/standard_blueprint.py
@@ -7,8 +7,9 @@ from app.common.file_type import FileType
 from app.schemas.document_request import DocumentRequest
 from app.schemas.success_code import SuccessCode
 from app.schemas.success_response import SuccessResponse
-from app.services.standard import processor
+from app.services.common.processor import preprocess_data
 from app.services.standard.vector_delete import delete_by_standard_id
+from app.services.standard.vector_store import vectorize_and_save
 
 standards = Blueprint('standards', __name__, url_prefix="/flask/standards")
 
@@ -23,7 +24,10 @@ def process_pdf_from_s3():
   status_code = HTTPStatus.OK
   try:
     if document_request.type == FileType.PDF:
-      status_code = processor.process_pdf(document_request)
+      chunks = preprocess_data(document_request)
+
+      # 5️⃣ 벡터화 + Qdrant 저장
+      vectorize_and_save(chunks, "standard", document_request)
   except Exception as e:
     raise e
